requirements.txt
audioop-lts==0.2.2
paho-mqtt==2.1.0
PyAudio==0.2.14
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0
